requirements.txt
# ── Telegram ──────────────────────────────────────────────────────────────────
telethon>=1.36

# ── LLM / Embeddings ──────────────────────────────────────────────────────────
openai>=1.30
httpx>=0.27          # TTS API calls (ElevenLabs + OpenAI)

# ── Vector store ──────────────────────────────────────────────────────────────
qdrant-client>=1.9
numpy>=1.26

# ── Storage ───────────────────────────────────────────────────────────────────
redis[hiredis]>=5.0  # Redis Streams event bus (async)

# ── API server ────────────────────────────────────────────────────────────────
fastapi>=0.111
uvicorn[standard]>=0.30

# ── Config / validation ───────────────────────────────────────────────────────
pydantic>=2.7
pydantic-settings>=2.3
python-dotenv>=1.0

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler>=3.10

# ── Optional: Claude provider ─────────────────────────────────────────────────
# anthropic>=0.28

# ── Optional: PostgreSQL ──────────────────────────────────────────────────────
# asyncpg>=0.29
# sqlalchemy[asyncio]>=2.0
